feat(ParamParsing): Updates to support parsing arbitrary parameters in route doc annotations
A large refactor of `Rolodex.Object`, including renaming it back
to `Rolodex.Schema`.

Rolodex.Schema still exposes the `schema/3` and `field/3` macros
for defining reusable parameter schemas. But, it also exposes a set
of functions for parsing arbitrary maps and keyword lists into maps
of structured parameter metadata. The macros use these functions to
define the schema shape. But, `Rolodex.Route` also uses these
functions to parse `@doc` annotations on controller actions.

With these changes, we can guarantee that shape of the metadata
sent to a `Rolodex.Processor` behaviour for every API route.

**Key Changes**

- `Rolodex.Object` -> `Rolodex.Schema`. Lots of additions to
  Rolodex.Schema as well for parsing parameters and traversing
  parameter maps for nested references to other schemas.
- Updates to `Rolodex.Route` to use the new `Rolodex.Schema.new_field/1`
  function to parse route headers, request bodies, query params, and
  path params.
- Updates to `Rolodex.generate_schemas/1` to traverse all parts of
  a parsed `%Route{}` for schema references
- Updates to `Rolodex.Processors.Swagger`: serialize parameters into
  valid JSON, restructure some things according to the JSON spec