Add support for response headers, reusable headers
We forgot to add support for defining response headers! That's the primary
addition this PR makes. In the process, we've also added support for defining
reusable headers via macros exported in the `Rolodex.Headers` module. The DSL
is very similar to what we've put together so far for schemas/responses/bodies.

Some highlights:

- You can now define shared headers via `Rolodex.Headers`
- Shared headers can be used in either responses OR route doc annotations (i.e.
  request headers)
- For OpenAPI 3, we never serialize headers as shared component refs b/c that
  isn't supported
- Removed documented support for nested schemas in parameters (path, query, header),
  since this isn't actually supported by OpenAPI 3. We weren't really testing against
  it anyway.