[PLATFORM-580] Support multiple response headers calls
Sometimes, you'll want to make multiple calls to the `headers/1` macro
when defining a response. This is often to use two or more shared
response headers modules for a specific response.

Via diff, we add support for multiple calls. All the headers passed
in to the response will ultimately be merged together when serialized.