[PLATFORM-616] Schemas for query and path params
This refactors our support to using schemas in the `query_params`
and `path_params` doc annotation items. Like with headers, we will
serialize these schemas in full in the Route parser. No changes are
needed to the way we serialize params in the OpenAPI processor, b/c
the data passed in will always be a map. I've chosen this approach,
rather than supporting serializing path and query params via component
references because of the slightly different shape of parameter descriptions.